+++ kofu/src/test/kotlin/org/springframework/fu/kofu/samples/cassandra.kt
@@ -2,9 +2,24 @@ package org.springframework.fu.kofu.samples
 
 import org.springframework.boot.WebApplicationType
 import org.springframework.fu.kofu.application
+import org.springframework.fu.kofu.cassandra.cassandra
 import org.springframework.fu.kofu.cassandra.reactiveCassandra
 
 fun cassandra() {
+	application(WebApplicationType.NONE) {
+		cassandra {
+			keyspaceName = "keyspaceOne"
+			clusterName = "cluster"
+			username = "user"
+			password = "password"
+			port = 9042
+			ssl = false
+			contactPoints.add("localhost")
+		}
+	}
+}
+
+fun reactiveCassandra() {
 	application(WebApplicationType.NONE) {
 		reactiveCassandra {
 			keyspaceName = "keyspaceOne"
@@ -17,4 +32,3 @@ fun cassandra() {
 		}
 	}
 }
-
